+++ repo-hooks-yarn/src/hooks/index.ts
@@ -2,3 +2,6 @@ export { useValidateToken } from "./useValidateToken";
 export { useGetPayloadFromToken } from "./useGetPayloadFromToken";
 export { useFetchUserById } from "./useFetchUserById";
 export { useDataProfileProcessor } from "./useDataProfileProcessor";
+export { useValidateRegisterForm } from "./useValidateRegisterForm";
+export { useValidateLoginForm } from "./useValidateLoginForm";
+export { useAuth } from "./useAuth";