+++ repo-hooks-yarn/CHANGELOG.md
@@ -9,14 +9,17 @@ El formato se basa en [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Cambiado
 
 - Se rellenó la información del README.
-
+- Se cambia el mapa de datos en UseDataProfileProcessor
 
 ## [0.1.2] - 2023-09-08
 
 ### Añadido
 
+- Se añade nuevo hook useAuth
 
 
+## [0.2.0] - 2023-09-08
+## [0.2.1] - 2023-09-08
 ### Cambiado